requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
effdet==0.4.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.2
packaging==25.0
pillow==11.3.0
protobuf==6.33.1
pycocotools==2.0.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
safetensors==0.7.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
timm==1.0.22
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typer-slim==0.20.0
typing_extensions==4.15.0
